feat(navigation) match sets to books, second commit
