+++ DPRHTML/extensions/src/components/navigationBookSelector.tsx
@@ -1,63 +1,90 @@
 import * as React from 'react';
 import { Grid, Typography, Select, Box } from '@material-ui/core';
+import { nikvoladi, G_kynames } from './navigation_common';
 
-const NavigationBookSelector = () => (
-  <Grid container direction="row">
-    <Grid item xs={6} container direction="row" alignItems="stretch">
-      <Grid item xs={12}>
-        <Typography variant="subtitle1" align="center">
-          <Box fontWeight="bold">Set</Box>
-        </Typography>
+const sets = ['Vinaya', 'Dīgha', 'Majjhima', 'Saṃyutta', 'Aṅguttara', 'Khuddaka', 'Abhidhamma'];
+const books = [nikvoladi['va'], nikvoladi['d'], nikvoladi['m'], nikvoladi['s'], nikvoladi['a'], G_kynames['k'], G_kynames['y']];
+var booksDisplayed = [];
+
+export default function NavigationBookSelector() {
+  const [setList, setSet] = React.useState<string[]>([]);
+  const [bookList, setBook] = React.useState<string[]>([]);
+
+  const handleChangeSet = (event: React.ChangeEvent<{ value: unknown }>) => {
+    const { options } = event.target as HTMLSelectElement;
+    const value: string[] = [];
+    for (let i = 0, l = options.length; i < l; i += 1) {
+      if (options[i].selected) {
+        value.push(options[i].value);
+        booksDisplayed = books[i];
+      }
+    }
+    setSet(value);
+  };
+
+  const handleChangeBook = (event: React.ChangeEvent<{ value: unknown }>) => {
+    const { options } = event.target as HTMLSelectElement;
+    const value: string[] = [];
+    for (let i = 0, l = options.length; i < l; i += 1) {
+      if (options[i].selected) {
+        value.push(options[i].value);
+      }
+    }
+    setBook(value);
+  };
+
+  return (
+    <Grid container direction="row">
+      <Grid item xs={6} container direction="row" alignItems="stretch">
+        <Grid item xs={12}>
+          <Typography variant="subtitle1" align="center">
+            <Box fontWeight="bold">Set</Box>
+          </Typography>
+        </Grid>
+        <Grid item xs={12}>
+          <Select
+            disableUnderline
+            style={{ width: '100%' }}
+            multiple
+            native
+            onChange={handleChangeSet}
+            inputProps={{
+              id: 'select-multiple-native',
+            }}
+          >
+            {sets.map(set => (
+              <option key={set} value={set}>
+                {set}
+              </option>
+            ))}
+          </Select>
+        </Grid>
       </Grid>
-      <Grid item xs={12}>
-        <Select
-          disableUnderline
-          style={{ width: '100%' }}
-          multiple
-          native
-          inputProps={{
-            id: 'select-multiple-native',
-          }}
-        >
-          <option value="v" label="Vinaya" />
-          <option value="d" label="Dīgha" />
-          <option value="m" label="Majjhima" />
-          <option value="s" label="Saṃyutta" />
-          <option value="a" label="Aṅguttara" />
-          <option value="k" label="Khuddaka" />
-          <option value="y" label="Abhidhamma" />
-        </Select>
+      <Grid item xs={6} container direction="row" alignItems="stretch">
+        <Grid item xs={12}>
+          <Typography variant="subtitle1" align="center">
+            <Box fontWeight="bold">Book</Box>
+          </Typography>
+        </Grid>
+        <Grid item xs={12}>
+          <Select
+            disableUnderline
+            style={{ width: '100%' }}
+            multiple
+            native
+            onChange={handleChangeBook}
+            inputProps={{
+              id: 'select-multiple-native',
+            }}
+          >
+            {booksDisplayed.map(book => (
+              <option key={book} value={book}>
+                {book}
+              </option>
+            ))}
+          </Select>
+        </Grid>
       </Grid>
     </Grid>
-    <Grid item xs={6} container direction="row" alignItems="stretch">
-      <Grid item xs={12}>
-        <Typography variant="subtitle1" align="center">
-          <Box fontWeight="bold">Book</Box>
-        </Typography>
-      </Grid>
-      <Grid item xs={12}>
-        <Select
-          disableUnderline
-          style={{ width: '100%' }}
-          multiple
-          native
-          inputProps={{
-            id: 'select-multiple-native',
-          }}
-        >
-          <option label="1" />
-          <option label="2" />
-          <option label="3" />
-          <option label="4" />
-          <option label="5" />
-          <option label="6" />
-          <option label="7" />
-          <option label="8" />
-          <option label="9" />
-        </Select>
-      </Grid>
-    </Grid>
-  </Grid>
-);
-
-export default NavigationBookSelector;
+  );
+}
